+++ server/internal/types/config/websocket.go
@@ -9,6 +9,7 @@ type WebSocket struct {
 	Password      string
 	AdminPassword string
 	Proxy         bool
+	Locks         []string
 }
 
 func (WebSocket) Init(cmd *cobra.Command) error {
@@ -27,6 +28,11 @@ func (WebSocket) Init(cmd *cobra.Command) error {
 		return err
 	}
 
+	cmd.PersistentFlags().StringSlice("locks", []string{}, "resources, that will be locked when starting (control, login)")
+	if err := viper.BindPFlag("locks", cmd.PersistentFlags().Lookup("locks")); err != nil {
+		return err
+	}
+
 	return nil
 }
 
@@ -34,4 +40,5 @@ func (s *WebSocket) Set() {
 	s.Password = viper.GetString("password")
 	s.AdminPassword = viper.GetString("password_admin")
 	s.Proxy = viper.GetBool("proxy")
+	s.Locks = viper.GetStringSlice("locks")
 }